+++ lib/zoi/types/number.ex
@@ -3,6 +3,10 @@ defmodule Zoi.Types.Number do
 
   def new(opts \\ []) do
     # Number is a union type of Float and Integer
-    Zoi.union([Zoi.integer(), Zoi.float()], opts)
+    {meta, opts} = Zoi.Types.Meta.create_meta(opts)
+
+    error = meta.error || "invalid number type"
+
+    Zoi.union([Zoi.integer(error: error), Zoi.float(error: error)], opts)
   end
 end